skill parser: extract CharacterSkills::Builder; slim facade
The heuristic core (slot/version building + role/trigger/progression inference)
moves into Builder, which owns @links/@version_keys and returns a graph with
links pre-filtered to built versions. CharacterSkillParser is now a thin façade
(~70 lines) wiring Builder/JpLocalizer/Reporter/Persister. Golden snapshot
unchanged — behavior preserved.
